+++ modules/core/dependency/python-ihm/ihm/reader.py
@@ -1055,6 +1055,10 @@ class _StructRefHandler(Handler):
         e = self.sysr.entities.get_by_id(entity_id)
         typ = self.type_map.get(db_name.lower())
         ref = self.sysr.references.get_by_id(id, typ)
+        # Strip newlines if code is split over multiple lines
+        if pdbx_seq_one_letter_code is not None:
+            pdbx_seq_one_letter_code \
+                = pdbx_seq_one_letter_code.replace('\n', '')
         self.copy_if_present(
             ref, locals(), keys=('db_name', 'db_code', 'details'),
             mapkeys={'pdbx_db_accession': 'accession',